backend/pyproject.toml
[tool.poetry]
name = "backend"        # Can be anything, doesn't have to match a folder
version = "0.1.0"
description = "..."
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langgraph = "^0.2.74"
langchain = "^0.3.19"
spacy = "^3.8.4"
typing-extensions = "^4.12.2"
langchain-openai = "^0.3.6"
python-dotenv = "^1.0.1"
pandas = "^2.0.2"
numpy = "^1.26.4"
elevenlabs = "^1.51.0"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
